parking-service/requirements.txt
black
flake8
flake8-bugbear
flake8-pie
isort
mypy
flask
SQLAlchemy
factory-boy
flask-sqlalchemy
pytest
